improving cleanup

+++ 10-web-test-main.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 set -e
-
+export DIR=$(dirname $0)
 export ION_HOST=$(test -z "$ION_HOST" && echo 'localhost:9090' || echo "$ION_HOST")
 export ION_ROOM=$(test -z "$ION_ROOM" && echo 'test-pink-video' || echo "$ION_ROOM")
 export HTTP_SCHEME=$(test -z "$ION_HTTP_SCHEME" && echo 'https' || echo "$ION_HTTP_SCHEME")
@@ -16,17 +16,18 @@ echo "Room: $ION_ROOM"
 echo
 echo "======================="
 echo
-echo "1. Joining $ION_ROOM with pink.video via go client"
+echo "1. Joining $ION_ROOM with pink.video via go client for up to 10 minutes"
 
-go run join.go -d 600 &
+(go run $DIR/join.go -d 600 | grep -v "DBG") &
 
 echo
 echo "======================="
 echo
 echo "2. Launching browser and searching for hot pink..."
-echo "This takes <= 20 seconds, plus browserstack queue time for MAC or IOS"
+echo "This takes ~ 20 seconds per browser, plus browserstack queue time for MAC or IOS"
+echo "Targets: $MULTI"
 
-/usr/bin/python3 browsertest.py
+/usr/bin/python3 $DIR/browsertest.py
 
 echo
 echo "======================="
@@ -34,6 +35,8 @@ echo
 echo "All done!"
 sleep 3
 PIDS=$(ps -ef | grep join | grep go-build | awk '{print $2}')
-echo "cleaning up... $PIDS"
+if [[ !-z "$PIDS" ]]; then
+    echo "cleaning up... $PIDS"
 
-kill -9 $PIDS
+    kill -9 $PIDS
+fi
